Full abstraction rework - step 1 : - ModuleData now has a direct reference to its ProtoPartModuleSnaopshot - PartData now has direct references to :   - The ProtoPartSnapshot   - The PartResource/ProtoPartResourceSnapshot list, through a common wrapper
This eliminate entirely the need for separate BackgroundUpdate/FixedUpdate methods, allowing to put the whole processing logic in ModuleData with permanent access to all stock objects, no matter the loaded/unloaded state.

The move of all resources to PartData as a common wrapper also allowed to significantly reduce the resource sim state specific code and subclassing mess.

The end goal is :
- To get ride of the Background class by allowing the creation of ModuleData derivatives for stock/mod modules.
- To get ride of PartModule config definitions and instead generalize a conceptually similar pattern as what we are doing with experiments.