Jobify realignment of body and vessel arrays between PositionCompute and FluxAnalysis
Don't return to main thread to reconstruct the arrays, do it in a job and pass the handle between.
Instead of passing a List of frames to FluxAnalysis, pass the component arrays and get back occlusion and irradiance arrays.
Remove unused Load/Save
Remove singleton pattern in SubstepSim, allow for multiple places to own their own copy and simulate independently as needed.