WIP: Jobify VesselData.EnvironmentUpdate
Demonstrate constructing a job and running it from the main thread.  Set up a job for the large summation across timesteps and bodies.  Not ideal, since we have to index and serialize each irradiance structure for each timestep, but demonstrative.  Potential for future improvements.
Switch FrameManager from Dictionary to LinkedList to improve performance (iteratating many sequential dictionary key lookups was noticably inefficient).
FIXMEs: Unbreak the variable-timestep awareness when summing, and the tracking of sunlightFactor and maybe visibility.  Move them into the supporting job calculation or replace.