Optimize Occlusion Test
Don't short-circuit the loop on occluders in the Job when occlusion is detected.  This is counter-intuitive, but empirically faster with Burst, perhaps because of removing the dependency on a result discovered at loop's end.
Avoid branching in occlusion test.
Since the actual distance doesn't matter, only if it is less than a target, we can skip the test for which end of the line segment is closest, and simply compute all 3 possibilities.  If any of them are sufficient, the full result is.  The branching is more expensive than the extra math operations.
Empirically on a 4C/4T 3.2GHz system with 20 vessels, 17 bodies, 67 block computations per FixedUpdate (100k warp, 30-sec intervals), this change reduced typical occlusion compute time from 1.7ms to to 1.2ms.